seek_hole: don't test filesystem-dependent behavior
Sadly, almost nothing about SEEK_HOLE and SEEK_DATA are well defined:

> A filesystem is not obliged to report holes, so these operations are
> not a guaranteed mechanism for mapping the storage space actually
> allocated to a file.  (Furthermore, a sequence of zeros that
> actually has been written to the underlying storage may not be
> reported as a hole.)  In the simplest implementation, a filesystem
> can support the operations by making SEEK_HOLE always return the
> offset of the end of the file, and making SEEK_DATA always return
> offset (i.e., even if the location referred to by offset is a hole,
> it can be considered to consist of data that is a sequence of
> zeros).

(from lseek(2))

This means that all we can say about SEEK_HOLE and SEEK_DATA when we
don't know what filesystem they're going to be used on is that they
won't return a value _less_ than offset if they succeed, and that
they'll always fail if offset is past the end of the file.  So that's
all we can test, or we'd have tests that failed on some filesystems
and not others.

This fixes these tests failing on ZFS for me (which partially, and not
reliably, implements SEEK_DATA and SEEK_HOLE, but in a way that's
still valid according to the man page).

Signed-off-by: Alyssa Ross <hi@alyssa.is>